reorganized config
