+++ src/commands/actions/goto.rs
@@ -1,13 +1,9 @@
-use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
-use std::str::FromStr;
 
 use crate::{
-	common::{
-		constraint::{Constraint, Criteria},
-		polybar, this_command,
-	},
-	i3::{self, get_current_output, get_filtered_workspaces},
+	common::{polybar, this_command},
+	config::global::GotoBehavior,
+	i3::{self, get_filtered_workspaces},
 };
 
 use crate::{CommandFn, Commands, DEFAULT_CMD, HELP_CMD, WILD_CMD};
@@ -21,41 +17,6 @@ lazy_static! {
 		cmds.insert(HELP_CMD, help as CommandFn);
 		cmds
 	};
-}
-
-#[derive(Debug, Deserialize, Serialize, PartialEq, Clone)]
-pub enum GotoBehavior {
-	Create,
-	Stop,
-}
-
-impl FromStr for GotoBehavior {
-	type Err = ();
-
-	fn from_str(s: &str) -> Result<Self, Self::Err> {
-		if s.len() == 0 {
-			return Ok(GotoBehavior::Stop);
-		}
-
-		match s.to_lowercase().as_ref() {
-			"create" => Ok(GotoBehavior::Create),
-			"stop" => Ok(GotoBehavior::Stop),
-			_ => {
-				eprintln!("Warning: Invalid value '{}' for 'navigation.goto.behavior'. Falling back to 'Stop'.", s);
-				Ok(GotoBehavior::Stop)
-			}
-		}
-	}
-}
-
-impl GotoBehavior {
-	#[allow(dead_code)]
-	pub fn to_string(&self) -> String {
-		match self {
-			GotoBehavior::Create => "create".to_string(),
-			GotoBehavior::Stop => "stop".to_string(),
-		}
-	}
 }
 
 pub fn help(_: Vec<String>) {
@@ -92,11 +53,7 @@ pub fn exec(mut args: Vec<String>) {
 		return;
 	}
 
-	let behavior: GotoBehavior = args
-		.get(0)
-		.unwrap_or(&"".to_string())
-		.parse()
-		.unwrap_or(GotoBehavior::Stop);
+	let behavior = GotoBehavior::from_argv(&mut args).unwrap();
 
 	let workspaces = get_filtered_workspaces(false);
 
@@ -105,7 +62,7 @@ pub fn exec(mut args: Vec<String>) {
 		if behavior == GotoBehavior::Create {
 			let last_ws = workspaces.first().unwrap();
 			i3::run_command(format!("workspace {}", last_ws.full_name()));
-			crate::commands::adjacent::exec(vec!["left".to_owned()]);
+			crate::commands::actions::adjacent::exec(vec!["left".to_owned()]);
 		}
 		return;
 	}
@@ -115,7 +72,7 @@ pub fn exec(mut args: Vec<String>) {
 			GotoBehavior::Create => {
 				let last_ws = workspaces.last().unwrap();
 				i3::run_command(format!("workspace {}", last_ws.full_name()));
-				crate::commands::adjacent::exec(vec!["right".to_owned()]);
+				crate::commands::actions::adjacent::exec(vec!["right".to_owned()]);
 			}
 			GotoBehavior::Stop => {}
 		};
