+++ src/commands/polybar/poke.rs
@@ -1,0 +1,5 @@
+use crate::common::polybar;
+
+pub fn exec(_: Vec<String>) {
+	polybar::update();
+}
